requirements.txt
pinecone-client 
streamlit 
langchain 
PyPDF2 
requests 
python-dotenv
